+++ src/crashplancli/extensions/__init__.py
@@ -8,7 +8,7 @@ from crashplancli.options import profile_option
 def sdk_options(f):
     """Decorator that adds two `click.option`s (--profile, --debug) to wrapped command, as well as
     passing the `crashplancli.options.CLIState` object using the [click.make_pass_decorator](https://click.palletsprojects.com/en/7.x/api/#click.make_pass_decorator),
-    which automatically instantiates the `pycpg` sdk using the crashplan profile provided from the `--profile`
+    which automatically instantiates the `pycpg` sdk using the CrashPlan profile provided from the `--profile`
     option. The `pycpg` sdk can be accessed from the `state.sdk` attribute.
 
     Example:
@@ -26,7 +26,7 @@ def sdk_options(f):
 
 
 script = ExtensionGroup(context_settings=CONTEXT_SETTINGS)
-"""A `click.Group` subclass that enables the crashplan CLI's custom error handling/logging to be used
+"""A `click.Group` subclass that enables the CrashPlan CLI's custom error handling/logging to be used
 in extension scripts. If only a single command is added to the `script` group it also uses that
 command as the default, so the command name doesn't need to be called explicitly.