+++ CHANGELOG.md
@@ -12,8 +12,13 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ## Updated
 
-- Removed support for python version 3.9 added support for python 3.13, includes many dependency updates
+- Increased python minimum version to 3.11
+- Added support for python 3.13, includes many dependency updates
+- Removed Incydr features and updated branding to CrashPlan
+- Updated legal-hold regardless of API client or username to use the same rest endpoints.
 
+## Changed
+ - Legal search requires a matter uid to search.
 
 ## Legacy Versions